feat: Add option to forward messages to a chat or user
Add the option to forward messages to a chat or user by specifying the alias, id, username, mobile, or id. This option can be used multiple times.

Refactor the `message` function in `telegram_upload/management.py` to include the new `forward` option. The function now accepts the `forward` option as a command-line argument and passes it to the `send_messages` method of the `TelegramManagerClient` class.

This commit also includes a minor fix to handle the `parse_mode` option when it is set to 'text'.